Release/v2.5.0

+++ cypress-tests/cypress.config.js
@@ -13,7 +13,7 @@ module.exports = defineConfig({
   requestTimeout: 10000,
   pageLoadTimeout: 20000,
   responseTimeout: 10000,
-  viewportWidth: 1200,
+  viewportWidth: 1440,
   viewportHeight: 960,
   chromeWebSecurity: false,
   trashAssetsBeforeRuns: true,
@@ -50,13 +50,17 @@ module.exports = defineConfig({
       on("task", {
         deleteFolder(folderName) {
           return new Promise((resolve, reject) => {
-            rmdir(folderName, { maxRetries: 10, recursive: true }, (err) => {
-              if (err) {
-                console.error(err);
-                return reject(err);
-              }
-              resolve(null);
-            });
+            if (fs.existsSync(folderName)) {
+              rmdir(folderName, { maxRetries: 10, recursive: true }, (err) => {
+                if (err) {
+                  console.error(err);
+                  return reject(err);
+                }
+                return resolve(null);
+              });
+            } else {
+              return resolve(null);
+            }
           });
         },
       });
@@ -73,7 +77,11 @@ module.exports = defineConfig({
     experimentalModfyObstructiveThirdPartyCode: true,
     experimentalRunAllSpecs: true,
     baseUrl: "http://localhost:8082",
-    specPattern: "cypress/e2e/**/*.cy.js",
+    specPattern: [
+      "cypress/e2e/editor/widget/*.cy.js",
+      "cypress/e2e/multipage/*.cy.js",
+    ],
+    downloadsFolder: "cypress/downloads",
     numTestsKeptInMemory: 25,
     redirectionLimit: 10,
     experimentalRunAllSpecs: true,
